example_openai_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "example_openai_agent"
version = "0.1.0"
description = "An example of a Synchronous Agentex Agent for use with OpenAI"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk>=0.5.3",
    "scale-gp",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88